UI: Extract shared item layout
