+++ packages/ui/src/menu/item.tsx
@@ -1,211 +1,15 @@
 import { Menu as _Menu } from '@base-ui/react/menu';
 import clsx from 'clsx';
-import {
-	Children,
-	cloneElement,
-	forwardRef,
-	isValidElement,
-	useId,
-} from '@wordpress/element';
-import { __, sprintf } from '@wordpress/i18n';
-import { VisuallyHidden } from '../visually-hidden';
+import { forwardRef } from '@wordpress/element';
 import resetStyles from '../utils/css/resets.module.css';
+import {
+	ItemLayout,
+	ItemLayoutContext,
+	useItemContent,
+} from '../utils/item-layout';
+import itemLayoutStyles from '../utils/item-layout/style.module.css';
 import styles from './style.module.css';
-import { MenuItemContentContext } from './context';
-import { ItemDescription } from './item-description';
-import { ItemLabel } from './item-label';
 import type { ItemProps } from './types';
-
-type ItemAriaProps = Pick<
-	ItemProps,
-	'aria-describedby' | 'aria-keyshortcuts' | 'aria-label' | 'aria-labelledby'
->;
-type UseItemContentOptions = ItemAriaProps & {
-	labelledBy?: string;
-	shortcut?: ItemProps[ 'shortcut' ];
-};
-
-function getStructuredItemContent( children: ItemProps[ 'children' ] ) {
-	const childArray = Children.toArray( children );
-	const label = childArray.find(
-		( child ) =>
-			isValidElement< { id?: string } >( child ) &&
-			child.type === ItemLabel
-	);
-	const description = childArray.find(
-		( child ) =>
-			isValidElement< { id?: string } >( child ) &&
-			child.type === ItemDescription
-	);
-
-	return {
-		descriptionId: isValidElement< { id?: string } >( description )
-			? description.props.id
-			: undefined,
-		hasDescription: !! description,
-		hasLabel: !! label,
-		hasStructuredContent: childArray.some(
-			( child ) =>
-				isValidElement( child ) &&
-				( child.type === ItemLabel || child.type === ItemDescription )
-		),
-		labelId: isValidElement< { id?: string } >( label )
-			? label.props.id
-			: undefined,
-	};
-}
-
-function useItemContent(
-	children: ItemProps[ 'children' ],
-	{
-		'aria-describedby': ariaDescribedBy,
-		'aria-keyshortcuts': ariaKeyShortcuts,
-		'aria-label': ariaLabel,
-		'aria-labelledby': ariaLabelledBy,
-		labelledBy: additionalLabelledBy,
-		shortcut,
-	}: UseItemContentOptions
-) {
-	const generatedLabelId = useId();
-	const generatedDescriptionId = useId();
-	const generatedShortcutDescriptionId = useId();
-	const {
-		descriptionId,
-		hasDescription,
-		hasLabel,
-		hasStructuredContent,
-		labelId,
-	} = getStructuredItemContent( children );
-	const resolvedLabelId =
-		labelId ??
-		( hasLabel || ! hasStructuredContent ? generatedLabelId : undefined );
-	const resolvedDescriptionId = descriptionId ?? generatedDescriptionId;
-	const shortcutDescriptionId = shortcut
-		? generatedShortcutDescriptionId
-		: undefined;
-
-	const describedBy = [
-		ariaDescribedBy,
-		hasDescription && resolvedDescriptionId,
-		shortcutDescriptionId,
-	]
-		.filter( Boolean )
-		.join( ' ' );
-	/*
-	 * `aria-labelledby` takes precedence over `aria-label` in the accessible
-	 * name algorithm. Only provide our generated label relationship when the
-	 * consumer has not supplied either explicit naming prop. Additional labels
-	 * are only appended to that generated relationship, so explicit naming stays
-	 * fully consumer-controlled.
-	 */
-	const labelledBy =
-		ariaLabelledBy ??
-		( ariaLabel
-			? undefined
-			: [ resolvedLabelId, additionalLabelledBy ]
-					.filter( Boolean )
-					.join( ' ' ) || undefined );
-
-	return {
-		contentContextValue: {
-			descriptionId: resolvedDescriptionId,
-			labelId: resolvedLabelId,
-		},
-		itemAriaProps: {
-			'aria-describedby': describedBy || undefined,
-			'aria-keyshortcuts': shortcut?.ariaKeyShortcut ?? ariaKeyShortcuts,
-			'aria-label': ariaLabel,
-			'aria-labelledby': labelledBy,
-		},
-		shortcutDescriptionId,
-	};
-}
-
-function ItemContent( {
-	children,
-	labelTrailing,
-	prefix,
-	shortcut,
-	shortcutDescriptionId,
-	suffix,
-	trailing,
-}: Pick< ItemProps, 'children' | 'prefix' | 'shortcut' | 'suffix' > & {
-	labelTrailing?: ItemProps[ 'suffix' ];
-	shortcutDescriptionId?: string;
-	trailing?: ItemProps[ 'suffix' ];
-} ) {
-	const hasStructuredContent =
-		getStructuredItemContent( children ).hasStructuredContent;
-	const itemChildren = hasStructuredContent ? (
-		Children.map( children, ( child ) => {
-			if (
-				isValidElement< { children?: ItemProps[ 'children' ] } >(
-					child
-				) &&
-				child.type === ItemLabel
-			) {
-				return cloneElement( child, {
-					children: (
-						<>
-							{ child.props.children }
-							{ labelTrailing }
-						</>
-					),
-				} );
-			}
-
-			return child;
-		} )
-	) : (
-		<ItemLabel>
-			{ children }
-			{ labelTrailing }
-		</ItemLabel>
-	);
-
-	return (
-		<>
-			{ prefix && (
-				<span className={ styles[ 'item-prefix' ] }>{ prefix }</span>
-			) }
-			<span className={ styles[ 'item-content' ] }>
-				<span className={ styles[ 'item-children' ] }>
-					{ itemChildren }
-				</span>
-				{ suffix && (
-					<span className={ styles[ 'item-suffix' ] }>
-						{ suffix }
-					</span>
-				) }
-				{ shortcut && (
-					<span
-						className={ styles[ 'item-shortcut' ] }
-						aria-hidden="true"
-					>
-						{ shortcut.displayShortcut }
-					</span>
-				) }
-				{ trailing && (
-					<span className={ styles[ 'item-trailing' ] }>
-						{ trailing }
-					</span>
-				) }
-			</span>
-			{ shortcut && shortcutDescriptionId && (
-				<VisuallyHidden
-					id={ shortcutDescriptionId }
-					render={ <span /> }
-				>
-					{ sprintf(
-						/* translators: %s: human-readable keyboard shortcut. */
-						__( 'Keyboard shortcut: %s' ),
-						shortcut.description
-					) }
-				</VisuallyHidden>
-			) }
-		</>
-	);
-}
 
 /**
  * Renders an individual menu item.
@@ -240,23 +44,24 @@ const Item = forwardRef< HTMLDivElement, ItemProps >( function MenuItem(
 			{ ...itemAriaProps }
 			className={ clsx(
 				resetStyles[ 'box-sizing' ],
+				itemLayoutStyles.item,
 				styles.item,
 				className
 			) }
 			{ ...props }
 		>
-			<MenuItemContentContext.Provider value={ contentContextValue }>
-				<ItemContent
+			<ItemLayoutContext.Provider value={ contentContextValue }>
+				<ItemLayout
 					prefix={ prefix }
 					shortcut={ shortcut }
 					shortcutDescriptionId={ shortcutDescriptionId }
 					suffix={ suffix }
 				>
 					{ children }
-				</ItemContent>
-			</MenuItemContentContext.Provider>
+				</ItemLayout>
+			</ItemLayoutContext.Provider>
 		</_Menu.Item>
 	);
 } );
 
-export { Item, ItemContent, useItemContent };
+export { Item };
