+++ packages/ui/src/menu/popup.tsx
@@ -1,8 +1,10 @@
 import { Menu as _Menu } from '@base-ui/react/menu';
 import clsx from 'clsx';
 import { forwardRef } from '@wordpress/element';
+import popupStyles from '../utils/css/dropdown-popup.module.css';
 import { renderSlotWithChildren } from '../utils/render-slot-with-children';
 import { ThemeProvider } from '../utils/theme-provider';
+import itemLayoutStyles from '../utils/item-layout/style.module.css';
 import styles from './style.module.css';
 import { Portal } from './portal';
 import { Positioner } from './positioner';
@@ -23,10 +25,11 @@ const Popup = forwardRef< HTMLDivElement, PopupProps >( function MenuPopup(
 			<_Menu.Popup
 				ref={ ref }
 				className={ clsx(
+					popupStyles.surface,
 					styles.popup,
 					isSubmenu
-						? styles[ 'popup--submenu' ]
-						: styles[ 'popup--root' ],
+						? popupStyles[ 'nested-motion' ]
+						: popupStyles[ 'root-motion' ],
 					className
 				) }
 				{ ...props }
@@ -36,7 +39,10 @@ const Popup = forwardRef< HTMLDivElement, PopupProps >( function MenuPopup(
 					 * `styles.list` provides the alignment scope for items
 					 * that are not inside an explicit Menu.Group.
 					 */
-					className={ styles.list }
+					className={ clsx(
+						itemLayoutStyles[ 'alignment-group' ],
+						styles.list
+					) }
 				>
 					{ children }
 				</div>
