+++ packages/ui/src/menu/radio-group.tsx
@@ -1,7 +1,7 @@
 import { Menu as _Menu } from '@base-ui/react/menu';
 import clsx from 'clsx';
 import { forwardRef } from '@wordpress/element';
-import styles from './style.module.css';
+import itemLayoutStyles from '../utils/item-layout/style.module.css';
 import type { RadioGroupProps } from './types';
 
 /**
@@ -12,7 +12,10 @@ const RadioGroup = forwardRef< HTMLDivElement, RadioGroupProps >(
 		return (
 			<_Menu.RadioGroup
 				ref={ ref }
-				className={ clsx( styles[ 'radio-group' ], className ) }
+				className={ clsx(
+					itemLayoutStyles[ 'layout-transparent' ],
+					className
+				) }
 				{ ...props }
 			/>
 		);
