+++ packages/ui/src/utils/item-layout/context.tsx
@@ -1,0 +1,14 @@
+import { createContext, useContext } from '@wordpress/element';
+
+export type ItemLayoutContextValue = {
+	labelId?: string;
+	descriptionId: string;
+};
+
+const ItemLayoutContext = createContext< ItemLayoutContextValue | null >(
+	null
+);
+
+const useItemLayoutContext = () => useContext( ItemLayoutContext );
+
+export { ItemLayoutContext, useItemLayoutContext };
